fix(install): make sure that a domain is configured even when the skip-ingress option is active
